Fixed duplicated gunicorn init logging message
The initial logging messages of gunicorn were duplicated (once with our
configured handler and once with the default handler).

To avoid this we directly pass the configuration to the wsgi application
as dict and we don't re-configure the logging within Flask.

For the `make serve` target which don't uses gunicorn, we added the
logging configuration directly within the entry point `service_alti.py`
which is only used in this case.