Fixed `make test` and `make dockerrun` on development pc
When running these targets on a development PC, a FileNotFoundError('/var/local/profile') was
raised. Instead of using '/var/local/profile' as default
DTM_BASE_PATH, we now use the current directory wich always
exist.

For the `make test` we also pass the makefile
DTM_BASE_PATH to avoid using the default path coded again to
'/var/local/profile'.

Made also the test a little bit more verbose.

The `dockerrun` target suffer also with the DTM_BASE_PATH, here instead
of passing the current repo dir  (which don't exist on the container) we
pass '.' which resolve to the current dir inside the container. Also
added the LOGGING_CFG in docker compose in order to be able to configure
another file.