Improved crash logging
When the application crashed, the log did not contains any back trace,
but only a stringification of the exception which can be very short and
unusefull. For example if a KeyError is raised, str(e) would return only
the name of the key that cause the exception, which such info it is
almost impossible to know what cause the exception.