Fixed initialization of raster files
The initialization did not happen because the way flask is started, the
main() method was never called ! Therefore removed the main() which
doesn't make sense at all in flask and changed the init raster logic.

In order to simplify the unittests, the raster preloading is
configurable now and by default disabled. On the deployement if we want
we can still enable preloading.