BGDIINF_SB-2197: Returns 415 in case of wrong content-type
If the request was done with an invalid content-type, then the service
returned a 400, "No 'geom' given, cannot create a profile without coordinates".

This was kind of misleading, for example when doing a request with curl without
specifying the content-type (curl used application/x-www-form-urlencoded) but
with valid payload.

The issue was that `request.is_json` in flask is set based on the content-type header
and not on the payload being a valid json.