BGDIINF_SB-2879: Added cache-control header
We experienced on our e2e tests issues with caching where we had unexpected
miss from cache. A first analyze showed us that the service did not set any
cache-control header, so in theory CF would use the default TTL of the cache
policy. Anyway it is better to control cache from the service and not from CF.